update and fix core
