+++ api/authentication/tests/__init__.py
@@ -1,0 +1,1 @@
+# This file makes authentication/tests a Python package
