+++ api/products/tests/__init__.py
@@ -1,0 +1,1 @@
+# This file makes products/tests a Python package
